templates/template/requirements.txt
tensorflow-gpu==1.1.0
